feat: add function to override tui files
GNOME has this annoying behavior of setting EVERYTHING it has as a
default, this leaks to other stuff like Sway. So when wofi runs in drun
mode it will open TUI applications using GNOME terminal and not Rio.
This script creates a function that copies the original desktop file,
changes it to NOT be a terminal (with Terminal=false) and then replaces
the Exec command with "rio -e <whatever>" so it FORCES the execution on
Rio (or any other terminal)